refactor: remove IConfigurationWire
